qcacmn: Change ce_service_max_yield_time ini to us granularity
Change ce_service_max_yield_time ini to us granularity to have
yield time of us order as well.

CRs-fixed: 2086190

+++ hif/inc/hif.h
@@ -789,7 +789,7 @@ hif_reg_based_get_target_info(struct hif_opaque_softc *hif_ctx,
  * Return: void
  */
 void hif_set_ce_service_max_yield_time(struct hif_opaque_softc *hif,
-				       uint8_t ce_service_max_yield_time);
+				       uint32_t ce_service_max_yield_time);
 
 /**
  * hif_get_ce_service_max_yield_time() - get CE service max yield time
